# Optimize typeof(T).IsValueType