--- a/mono/mini/method-to-ir.c
+++ b/mono/mini/method-to-ir.c
@@ -10062,8 +10062,7 @@ field_access_end:
 						MonoMethod *is_vt_method = mini_get_method (cfg, method, is_vt_token, NULL, generic_context);
 						if (is_vt_method->klass == mono_defaults.systemtype_class &&
 							!mini_is_gsharedvt_variable_klass (tclass) &&
-							!mono_class_is_open_constructed_type ((m_class_get_byval_arg (tclass)) &&
-							!mono_class_is_ginst (klass) &&
+							!mono_class_is_open_constructed_type (m_class_get_byval_arg (tclass)) &&
 							!strcmp ("get_IsValueType", is_vt_method->name)) {
 							next_ip = is_vt_ip;
 							EMIT_NEW_ICONST (cfg, ins, m_class_is_valuetype (tclass) ? 1 : 0);