Optimize `typeof(T).IsValueType` to just `true` of `false` (I am porting my fix for CoreCLR: https://github.com/dotnet/runtime/pull/1157 to mono). E.g.:
```csharp
static void Validate<T>()
{
    if (!typeof(T).IsValueType)
        throw new Exception("Not allowed");
}
```
Current codegen for `Validate<int>()`:
```asm
0000000000000000 <gram_Validate_int____>:
<BB>:1
   0:   50                      push   %rax
   1:   48 b8 40 22 25 01 00    movabs $0x1252240,%rax
   8:   00 00 00
   b:   48 bf 30 cb 7d d4 bf    movabs $0x7fbfd47dcb30,%rdi
  12:   7f 00 00
  15:   ff 10                   callq  *(%rax)
  17:   84 c0                   test   %al,%al
  19:   74 02                   je     1d <gram_Validate_int____+0x1d>
  1b:   58                      pop    %rax
  1c:   c3                      retq
  1d:   48 b8 48 22 25 01 00    movabs $0x1252248,%rax
  24:   00 00 00
  27:   bf 70 2e f3 00          mov    $0xf32e70,%edi
  2c:   be 01 00 00 00          mov    $0x1,%esi
  31:   ff 10                   callq  *(%rax)
  33:   48 b9 50 22 25 01 00    movabs $0x1252250,%rcx
  3a:   00 00 00
  3d:   bf 17 00 00 02          mov    $0x2000017,%edi
  42:   48 89 c6                mov    %rax,%rsi
  45:   ff 11                   callq  *(%rcx)
  47:   48 b9 58 22 25 01 00    movabs $0x1252258,%rcx
  4e:   00 00 00
  51:   48 89 c7                mov    %rax,%rdi
  54:   ff 11                   callq  *(%rcx)
```
New codegen for `Validate<int>()`:
```asm
0000000000000000 <gram_Validate_int____>:
   0:   c3                      retq
```

Should I ignore gsharedvt types here?